test01에 있던 window, Tablewidget 두가지를 Mainwindow.py와 Tablewidget 두가지로 나누어 구현하였습니다. Tablewidget에 Printer 다이얼로그를 추가하였습니다. 정상작동되는지는 차후 확인해보아야합니다.
