+++ test02.py
@@ -5,7 +5,7 @@ import tkinter as tk
 import tkinter.messagebox as tmsg
 from tkinter import *
 import tkinter.filedialog as fdialog
-from test01 import MyMainWindow
+from Mainwindow import MyMainWindow
 import sys
 from PyQt5.QtWidgets import *
