--- test03.py
@@ -1,32 +1,0 @@
-import sys
-from PyQt5.QtWidgets import (QApplication, QWidget, QGridLayout, QLabel, QLineEdit, QTextEdit)
-
-
-class MyWindow(QWidget):
-
-    def __init__(self):
-        super().__init__()
-        self.initUI()
-
-    def initUI(self):
-        grid = QGridLayout()
-        self.setLayout(grid)
-
-        grid.addWidget(QLabel('Title:'), 0, 0)
-        grid.addWidget(QLabel('Author:'), 1, 0)
-        grid.addWidget(QLabel('Review:'), 2, 0)
-
-        grid.addWidget(QLineEdit(), 0, 1)
-        grid.addWidget(QLineEdit(), 1, 1)
-        grid.addWidget(QTextEdit(), 2, 1)
-
-        self.setWindowTitle('QGridLayout')
-        self.setGeometry(500, 500, 300, 200)
-
-        self.show()
-
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    ex = MyWindow()
-    sys.exit(app.exec_())
